organized the file directory
