+++ src/assets/styles/main.scss
@@ -4,8 +4,8 @@
 
 @import "theme/theme";
 
-@import "areas/reading-area";
+@import "page/reading-area";
 
-@import "areas/booklist";
+@import "page/booklist";
 
 @import "page/authors";
