+++ src/assets/styles/page/_booklist.scss
@@ -5,7 +5,7 @@
 }
 .book-title{
   color: #2d3235;
-  font-weight: 550;
+  font-weight: 500;
 }
 
 .add-book-title{
